de-associate hosts from streaming threads when both the receiver and sender stop, so that each time the threads are re-balanced

+++ src/database/rrd.h
@@ -1181,7 +1181,7 @@ struct rrdhost {
 
     struct {
         struct stream_thread *thread;
-        uint16_t refcount;
+        uint8_t refcount;
 
         // --- sender ---
